bug fix: regex failed to load due to circular dependency
